change from Long to Integer

+++ zcms/src/com/mavict/account/AccountController.java
@@ -54,7 +54,7 @@ public class AccountController extends BaseController{
 	
 	/** 保存 */
 	@RequestMapping("/save")
-	public String save(RedirectAttributes redirectAttributes,Account account, Boolean isEnabled, Long accountRoleId){
+	public String save(RedirectAttributes redirectAttributes,Account account, Boolean isEnabled, Integer accountRoleId){
 		account.setIsEnabled(isEnabled);
 		account.setIsLocked(false);
 		account.setLoginCount(0);
@@ -81,7 +81,7 @@ public class AccountController extends BaseController{
 	
 	/** 编辑 */
 	@RequestMapping("/edit/{id}")
-	public String edit(@PathVariable Long id, ModelMap model){
+	public String edit(@PathVariable Integer id, ModelMap model){
 		model.addAttribute("account", accountService.getService(id));
 		model.addAttribute("roles", accountRoleService.getAllService());
 		return "/admin/account/edit";
@@ -89,7 +89,7 @@ public class AccountController extends BaseController{
 	
 	/** 更新 */
 	@RequestMapping("/update")
-	public String update(RedirectAttributes redirectAttributes,Account account, String newPassword, Long roleId, boolean isEnabled, boolean isLocked,HttpServletRequest request){
+	public String update(RedirectAttributes redirectAttributes,Account account, String newPassword, Integer roleId, boolean isEnabled, boolean isLocked,HttpServletRequest request){
 		/** 密码表单不为空,重置密码 */
 		if (!(newPassword == null ||"".equals(newPassword))) {
 			account.setPassword(EncryptUtils.encrypt(newPassword));
@@ -132,7 +132,7 @@ public class AccountController extends BaseController{
 		
 //		accountService.getAccountService("admin");
 		
-		accountService.getService(1L);
+		accountService.getService(1);
 		
 //		Account account = new Account();
 //		account.setUsername("zhangtao");
