+++ zcms/src/com/mavict/article/image/ArticleImageDao.java
@@ -11,7 +11,7 @@ import com.mavict.base.BaseDao;
  * @author 沧海软件(北京)有限公司
  * @date 2016年1月16日 上午10:55:21
  */
-public interface ArticleImageDao extends BaseDao<ArticleImage, Long> {
+public interface ArticleImageDao extends BaseDao<ArticleImage, Integer> {
 
 	/**
 	 * 获取Article上传文章图片
