+++ zcms/src/com/mavict/setting/navigation/NavigationDao.java
@@ -11,12 +11,12 @@ import com.mavict.base.BaseDao;
  * @date   2015年8月2日 上午10:10:59
  */
 
-public interface NavigationDao extends BaseDao<Navigation, Long> {
+public interface NavigationDao extends BaseDao<Navigation, Integer> {
 	/** 获取所有分类  */
 	List<Navigation> getAll();
 	
 	/** 判断是否有下级栏目*/
-	boolean hasChild(Long parentId);
+	boolean hasChild(Integer parentId);
 	
 
 }
